+++ cypress/integration/WidgetsForEnrollmentPages/WidgetsForEnrollmentEditEvent/index.js
@@ -1,0 +1,7 @@
+import '../sharedSteps';
+import '../WidgetEnrollment';
+import '../WidgetProfile';
+
+beforeEach(() => {
+    cy.loginThroughForm();
+});