+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/EnrollmentPageDefault.container.js
@@ -1,14 +1,20 @@
 // @flow
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState, useEffect } from 'react';
+import log from 'loglevel';
+import { errorCreator } from 'capture-core-utils';
 // $FlowFixMe
 import { useSelector, shallowEqual, useDispatch } from 'react-redux';
-import { useHistory } from 'react-router';
-import log from 'loglevel';
 import { useEnrollment } from 'capture-core/hooks/useEnrollment';
-import { errorCreator } from 'capture-core-utils';
+import { useHistory } from 'react-router';
 import { useProgramInfo } from '../../../../hooks/useProgramInfo';
 import { EnrollmentPageDefaultComponent } from './EnrollmentPageDefault.component';
-import { useTeiAttributes, useProgramMetadata } from './hooks';
+import {
+    useEnrollment,
+    useTeiAttributes,
+    useProgramMetadata,
+    useHideWidgetByRuleLocations,
+    useProgramStages,
+} from './hooks';
 import { runRulesForEnrollment } from './runRulesForEnrollment';
 import { urlArguments } from '../../../../utils/url';
 import { deleteEnrollment } from '../EnrollmentPage.actions';
@@ -37,9 +43,12 @@ export const EnrollmentPageDefault = () => {
     const { error: teiAttributesError, attributes } = useTeiAttributes(teiId);
     const { error: enrollmentsError, enrollment } = useEnrollment(teiId);
     const { error: programMetaDataError, programMetadata } = useProgramMetadata(programId);
+    const stages = useProgramStages(program, programMetadata.programStages);
 
     if (programMetaDataError || enrollmentsError || teiAttributesError) {
-        log.error(errorCreator('Enrollment page could not be loaded')({ programMetaDataError, enrollmentsError, teiAttributesError }));
+        log.error(errorCreator('Enrollment page could not be loaded')(
+            { programMetaDataError, enrollmentsError, teiAttributesError },
+        ));
     }
 
     const [ruleEffects, setRuleEffects] = useState(undefined);
@@ -54,16 +63,7 @@ export const EnrollmentPageDefault = () => {
         }
     }, [orgUnit, program, programMetadata, enrollment, attributes]);
 
-    const flatRuleActionLocations = useMemo(() => program.programRules.map(item => item.programRuleActions
-        .map(rule => rule.location || null))
-        .flat(), [program.programRules]);
-
-    const hideWidgets = useMemo(() => {
-        const hideWidgetObject = {};
-        hideWidgetObject.feedback = !flatRuleActionLocations.includes('feedback');
-        hideWidgetObject.indicator = !flatRuleActionLocations.includes('indicators');
-        return hideWidgetObject;
-    }, [flatRuleActionLocations]);
+    const hideWidgets = useHideWidgetByRuleLocations(program.programRules);
 
 
     const onDelete = () => {
@@ -77,6 +77,8 @@ export const EnrollmentPageDefault = () => {
         <EnrollmentPageDefaultComponent
             teiId={teiId}
             program={program}
+            stages={stages}
+            events={enrollment?.events ?? []}
             enrollmentId={enrollmentId}
             onDelete={onDelete}
             widgetEffects={outputEffects}