+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/hooks/index.js
@@ -1,4 +1,8 @@
 import { useTeiAttributes } from './useTeiAttributes';
 import { useProgramMetadata } from './useProgramMetadata';
+import { useHideWidgetByRuleLocations } from './useHideWidgetByRuleLocations';
+import { useProgramStages } from './useProgramStages';
+
+export { useEnrollment, useProgramMetadata, useTeiAttributes, useHideWidgetByRuleLocations, useProgramStages };
 
 export { useProgramMetadata, useTeiAttributes };