+++ src/core_modules/capture-core/metaData/index.js
@@ -7,6 +7,8 @@ export {
     dataElementTypes,
     Unique as DataElementUnique,
     uniqueScope as dataElementUniqueScope,
+    statusTypes,
+    translatedStatusTypes,
 } from './DataElement';
 export { Icon } from './Icon';
 export {
@@ -22,7 +24,6 @@ export { SearchGroup } from './SearchGroup';
 export { InputSearchGroup } from './InputSearchGroup';
 export { TrackedEntityType, TeiRegistration } from './TrackedEntityType';
 export { SystemSettings } from './SystemSettings';
-
 // helpers
 export {
     getScopeFromScopeId,